all code refactoring 13.11.2019
