+++ src/main/java/ru/house/manager/BLDB/Util.java
@@ -1,4 +1,4 @@
-package BLBD;
+package ru.house.manager.BLDB;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
