+++ src/main/java/ru/house/manager/Hash/MagicString/magicString.java
@@ -1,4 +1,4 @@
-package MagicString;
+package ru.house.manager.Hash.MagicString;
 
 public class magicString {
     public static final String MAGIC_STRING="sdjDS8721Ssfxc";
